Add BS to python

+++ python/src/_python/types.h
@@ -2,9 +2,10 @@
 #define _MLIB_PYTHON_TYPES_
 
 #include <cstddef>
+#include <cstdint>
 
 /// ----------------------------------------------------------------------------
 /// type used for intergers
-using MLIBInt = std::size_t;
+using MLIBInt = std::uint64_t;
 
 #endif // _MLIB_PYTHON_TYPES_
